finished test project modules implementation

+++ app/Http/Controllers/Admin/TestProjectModuleController.php
@@ -6,11 +6,13 @@ use App\Http\Controllers\Controller;
 use App\Models\TestProjects\TestProject;
 use App\Models\TestProjects\TestProjectModule;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TestProjectModuleController extends Controller
 {
     public function create(TestProject $testProject)
     {
+        Gate::authorize('create', TestProjectModule::class);
         return view('admin.test-projects.modules.create', [
             'testProject' => $testProject
         ]);
@@ -21,6 +23,7 @@ class TestProjectModuleController extends Controller
         TestProject $testProject
     )
     {
+        Gate::authorize('create', TestProjectModule::class);
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string|max:255',
@@ -47,6 +50,7 @@ class TestProjectModuleController extends Controller
         TestProjectModule $module
     )
     {
+        Gate::authorize('update', $module);
         return view('admin.test-projects.modules.edit', [
             'testProject' => $testProject,
             'module' => $module,
@@ -59,6 +63,7 @@ class TestProjectModuleController extends Controller
         TestProjectModule $module
     )
     {
+        Gate::authorize('update', $module);
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string|max:255',
@@ -86,6 +91,7 @@ class TestProjectModuleController extends Controller
         TestProjectModule $module
     )
     {
+        Gate::authorize('delete', $module);
         return view('admin.test-projects.modules.delete', [
             'testProject' => $testProject,
             'module' => $module,
@@ -97,6 +103,7 @@ class TestProjectModuleController extends Controller
         TestProjectModule $module
     )
     {
+        Gate::authorize('delete', $module);
         $module->delete();
         return redirect()->route("admin.test-projects.show", [
             "test_project" => $testProject,
